+++ src/pages/TriageOverdue.vue
@@ -1,5 +1,4 @@
 <script setup>
 </script>
-
 <template>
 </template>
